Extract motion event handling into controller
